JonathanExperimental update from main

+++ src/main/java/be/intec/themarujohyperblog/service/PostServiceImpl.java
@@ -145,12 +145,13 @@ public class PostServiceImpl implements PostService {
     }
 
 
+
     public long countAllBlogPosts() {
         return (int) postRepository.count();
     }
 
 
-    public int countPosts() {
+    public long countPosts() {
         return (int) postRepository.count();
     }
